:art: refactor post
