+++ lib/views/posts/post_detail_view.dart
@@ -1,7 +1,7 @@
 import 'package:ecogest_front/state_management/posts/posts_cubit.dart';
 import 'package:ecogest_front/state_management/posts/posts_state.dart';
 import 'package:ecogest_front/widgets/app_bar.dart';
-import 'package:ecogest_front/widgets/post_/single_post.dart';
+import 'package:ecogest_front/widgets/post/single_post.dart';
 import 'package:flutter/material.dart';
 import 'package:ecogest_front/widgets/bottom_bar.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -41,7 +41,9 @@ class _PostDetailView extends State<PostDetailView> {
                 child: Text(state.message),
               );
             } else if (state is OnePostStateSuccess) {
-              return SinglePostWidget(post: state.post!);
+              return SingleChildScrollView(
+                child: SinglePostWidget(post: state.post!),
+              );
             }
             return const SizedBox.shrink();
           },
