+++ lib/widgets/post/post_body/post_body_stats_category_icon.dart
