+++ lib/widgets/post/posts_list.dart
@@ -1,16 +1,11 @@
 import 'package:ecogest_front/models/post_model.dart';
 import 'package:ecogest_front/models/user_model.dart';
 import 'package:ecogest_front/state_management/authentication/authentication_cubit.dart';
-import 'package:ecogest_front/state_management/like/like_cubit.dart';
 import 'package:ecogest_front/state_management/posts/posts_cubit.dart';
-import 'package:ecogest_front/widgets/post/post_content_author.dart';
-import 'package:ecogest_front/widgets/post/post_content_buttons.dart';
-import 'package:ecogest_front/widgets/post/post_content_infos.dart';
-import 'package:ecogest_front/widgets/post/post_separator.dart';
+import 'package:ecogest_front/widgets/post/single_post.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:ecogest_front/assets/ecogest_theme.dart';
 
 class PostsList extends StatefulWidget {
   PostsList(
@@ -65,7 +60,6 @@ class _PostsList extends State<PostsList> {
     final List<PostModel> posts = widget.posts;
     final bool isLastPage = widget.isLastPage;
     final Function() onScrolled = widget.onScrolled;
-    final UserModel? user = context.read<AuthenticationCubit>().state.user;
 
     return NotificationListener(
       child: RefreshIndicator(
@@ -80,46 +74,14 @@ class _PostsList extends State<PostsList> {
           itemCount: posts.length + (isLastPage ? 1 : 0),
           itemBuilder: (BuildContext context, int index) {
             if (index < posts.length) {
-              return Container(
-               child: Column(
-                  children: [
-                    // Author info
-                    PostContentAuthor(
-                        author: posts[index].user,
-                        position: posts[index].position,
-                        date: posts[index].createdAt,
-                        post: posts[index]),
-                    const PostSeparator(),
-                    // Post info
-                    InkWell(
+              return InkWell(
                       onTap: () {
                         // Redirect to post detail page
                         GoRouter.of(context).push('/posts/${posts[index].id!}');
                       },
-                      child: PostContentInfos(post: posts[index]),
-                    ),
-                    // Buttons
-                    BlocProvider<LikeCubit>(
-                      create: (context) => LikeCubit(),
-                      child: PostContentButtons(
-                        post: posts[index],
-                        likes: posts[index].likes!.length,
-                        isLiked: posts[index]
-                                .likes
-                                ?.any((like) => like.userId == user!.id) ??
-                            false, // TODO
-                        comments: posts[index].comments,
-                        isChallenge:
-                            (posts[index].type.toString() == 'challenge')
-                                ? true
-                                : false,
-                        canEndChallenge: canEndChallenge(posts[index]),
-                      ),
-                    ),
-                  ],
-                ),
-              );
-            } else if (isLastPage) {
+                      child: SinglePostWidget(post: posts[index]),
+                    );
+                          } else if (isLastPage) {
               return const Center(
                 child: Text("Pas de nouvelles publications à afficher"),
               );
